Update exercise11_2.py
The original solution had a traceback error, as either the sum of numbers or count was not being added properly, which caused the program to return a division by zero error. This solution outputs the exact value as required per the textbook, and does not require you to create an additional list.